Update and rename PrescriptionsDB Task 2.sql to PrescriptionsDB Queries .sql
The first stage of your task is to create a database and import the three tables from the csv file. You should also add the necessary primary and foreign key constraints to the tables and provide a database diagram in your report which shows the three tables and their relationships. You should create the database with the name PrescriptionsDB and the tables with the following names:
a. Medical_Practice
b. Drugs
c. Prescriptions
You should also leave the column names as they appear in the csv file. This is so we can re-run your code.
2. Write a query that returns details of all drugs which are in the form of tablets or capsules. You can assume that all drugs in this form will have one of these words in the BNF_DESCRIPTION column.
3. Write a query that returns the total quantity for each of prescriptions – this is given by the number of items multiplied by the quantity. Some of the quantities are not integer values and your client has asked you to round the result to the nearest integer value.
4. Write a query that returns a list of the distinct chemical substances which appear in the Drugs table (the chemical substance is listed in the CHEMICAL_SUBSTANCE_BNF_DESCR column)